+++ bluetti_mqtt/server_cli.py
@@ -104,7 +104,8 @@ class CommandLineHandler:
             parser.print_help()
 
     def start(self, args: argparse.Namespace):
-        loop = asyncio.get_event_loop()
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
 
         # Register signal handlers for safe shutdown
         if sys.platform != 'win32':
@@ -119,6 +120,7 @@ class CommandLineHandler:
             loop.create_task(self.run(args))
             loop.run_forever()
         finally:
+            asyncio.set_event_loop(None)
             loop.close()
             logging.debug("Shut down completed")